Added functions in DBService to create, read, update, and delete items. Also abstracted functions from NotesController, though errors remain because unsure of whether the return types of some functions should be void or Task<ActionResult<string>>
